Moved SchemaLocator to its own file
