+++ src/native/etw_types/event_record.rs
@@ -107,7 +107,7 @@ impl EventRecord {
     /// # Example
     /// ```
     /// # use ferrisetw::native::etw_types::EventRecord;
-    /// # use ferrisetw::schema::SchemaLocator;
+    /// # use ferrisetw::schema_locator::SchemaLocator;
     /// use windows::Win32::System::Diagnostics::Etw::EVENT_HEADER_EXT_TYPE_RELATED_ACTIVITYID;
     ///
     /// let my_callback = |record: &EventRecord, schema_locator: &mut SchemaLocator| {
